Extract controller, align partials
