+++ app/views/schools/advice/electricity_meter_breakdown/_analysis.html.erb
@@ -6,7 +6,9 @@
        t('advice_pages.electricity_long_term.charts.group_by_week_electricity_meter_breakdown_one_year.subtitle')
      end %>
   <% c.with_header do %>
-    <p><%= t('advice_pages.electricity_long_term.charts.group_by_week_electricity_meter_breakdown_one_year.header') %></p>
+    <p>
+      <%= t('advice_pages.electricity_long_term.charts.group_by_week_electricity_meter_breakdown_one_year.header') %>
+    </p>
   <% end %>
 <% end %>
