+++ app/views/schools/advice/electricity_meter_breakdown/_insights.html.erb
@@ -1,9 +1,9 @@
 <%= component 'notice', status: :neutral do |c| %>
   <% c.with_link do
-       link_to t('advice_pages.electricity_meter_breakdown.insights.what_is_a_meter_breakdown.link'),
+       link_to t("advice_pages.#{@advice_page.fuel_type}_meter_breakdown.insights.what_is_a_meter_breakdown.link"),
                learn_more_school_advice_electricity_meter_breakdown_path(@school)
      end %>
-  <%= t('advice_pages.electricity_meter_breakdown.insights.what_is_a_meter_breakdown.text_html') %>
+  <%= t("advice_pages.#{@advice_page.fuel_type}_meter_breakdown.insights.what_is_a_meter_breakdown.text_html") %>
 <% end %>
 
 <%= render 'schools/advice/section_title',
